Made code more modular by having createListingCard in another file, and also destructured and included JSDOC comments.
